requirements.txt
boto3==1.17.106
aioboto3==9.2.0
qrcode
pillow
